Author and book command unit test has been done

+++ BookStore/Applications/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -4,9 +4,9 @@ namespace BookStore.Applications.AuthorOperations.Commands.DeleteAuthor
 {
     public class DeleteAuthorCommand
     {
-        private readonly BookStoreDbContext _context;
+        private readonly IBookStoreDbContext _context;
         public int Id { get; set; }
-        public DeleteAuthorCommand(BookStoreDbContext context)
+        public DeleteAuthorCommand(IBookStoreDbContext context)
         {
             _context = context;
         }
@@ -15,6 +15,8 @@ namespace BookStore.Applications.AuthorOperations.Commands.DeleteAuthor
             var result = _context.Authors.FirstOrDefault(a => a.Id == Id);
             if (result is null)
                 throw new InvalidOperationException("Author Bulunamadı");
+            if (_context.Books.FirstOrDefault(a=>a.AuthorId==result.Id)!=null)
+                throw new InvalidOperationException("Kullanıcı silinemez kayıtlı kitap var");
             _context.Authors.Remove(result);
             _context.SaveChanges();
         }
